Enhance signal quality and position management
- OBI persistence: wall-clock timer immune to exchange drift
- Momentum filter: venue-specific routing (B impulses check tracker_b)
- Per-symbol isolation: 8 independent engines prevent cross-symbol bleed
- OBI divergence: trending condition (both positive, one stronger)
- Impulse sizing: position scales 0.5-2x with impulse magnitude
- Exit timeout: 30s -> 5s, entry threshold 5 -> 5.5 bps
- Update documentation: HANDOVER.md v0.1.5, architecture.md, module_signal.md
- All tests pass

+++ src/config/schema.rs
@@ -207,8 +207,8 @@ pub struct StrategySettings {
     pub venue_freshness_ms: u64,
 
     /// Minimum cross-venue edge in bps to enter a trade (should cover fees + slippage)
-    #[validate(range(min = 2, max = 50))]
-    pub entry_threshold_bps: u64,
+    #[validate(range(min = 0.5, max = 50.0))]
+    pub entry_threshold_bps: f64,
 
     /// Cooldown in ms between trades for same (symbol, side) pair
     #[validate(range(min = 5, max = 5000))]
